build: Introduce android-info-extra.txt support
This allows adding android-info-extra.txt to OTA packaging folder
in case it exists in the target device tree, so that it can then
be used in conjunction with custom releasetools. Historically, devices
used to do assertions based on the modem version or even trusted zone
version in order to prevent flashing builds if certain conditions were
not met. That however, breaks `fastboot flashall` command since it looks
for the requirements set in board-info.txt.

With this, targets can define whatever non-standard requirements
in the newly introduced android-info extra file, keeping OTA assertions
in place and `fastboot flashall` command working without having to pass
`--force` argument.

Change-Id: I029d778b44be4d9d877701d05124451052a4b73c
Signed-off-by: Hadad <hadad@linuxmail.org>